+++ extensions/tidyread---streamline-your-daily-reading/src/add-source.command.tsx
@@ -1,13 +1,5 @@
-import { LaunchType, launchCommand } from "@raycast/api";
-import SourceForm from "./components/SourceForm";
+import RSSSearch from "./components/RSSSearch";
 
 export default function CreateSourceForm() {
-  return (
-    <SourceForm
-      onSuccess={async () => {
-        // don't have to handle error here, because it will be handled in SourceForm
-        await launchCommand({ name: "manage-source-list.command", type: LaunchType.UserInitiated });
-      }}
-    />
-  );
+  return <RSSSearch />;
 }