+++ extensions/tidyread---streamline-your-daily-reading/src/components/SourceForm.tsx
@@ -128,10 +128,10 @@ function InnerReadForm(props: { defaultValues?: Source; navigationTitle?: string
 
   return (
     <Form
-      searchBarAccessory={<Form.LinkAccessory target="https://tidyread.info/docs/where-to-find-rss" text="Find RSS" />}
+      searchBarAccessory={<Form.LinkAccessory target="https://tidyread.info/docs/where-to-find-rss" text="🤔 How To Find RSS?" />}
       actions={
         <CustomActionPanel>
-          <Action.SubmitForm icon={Icon.SaveDocument} title="Save Source" onSubmit={handleSubmit} />
+          <Action.SubmitForm icon={Icon.SaveDocument} title="Save" onSubmit={handleSubmit} />
         </CustomActionPanel>
       }
       navigationTitle={navigationTitle}
@@ -157,8 +157,8 @@ function InnerReadForm(props: { defaultValues?: Source; navigationTitle?: string
       <Form.TextField
         {...itemProps.rssLink}
         title="RSS Link"
-        placeholder="Enter RSS link (optional)"
-        info="Used for generate digest, you can learn how to find rss through the documentation in the upper right corner"
+        placeholder="Enter RSS link for AI digest (optional)"
+        info="Used for generating digest. You can learn how to find rss through the documentation in the upper right corner."
       />
       {!!values.rssLink && (
         <Form.TextField
@@ -170,7 +170,7 @@ function InnerReadForm(props: { defaultValues?: Source; navigationTitle?: string
       )}
       <Form.TagPicker {...itemProps.tags} title="Tags" placeholder="Select tags (optional)">
         {CATEGORIES.map((category) => (
-          <Form.TagPicker.Item key={category} value={category} title={category} />
+          <Form.TagPicker.Item key={category.value} value={category.value} title={category.value} />
         ))}
       </Form.TagPicker>
     </Form>