+++ extensions/tidyread---streamline-your-daily-reading/src/digest/index.ts
@@ -19,7 +19,7 @@ import { addUtmSourceToUrl } from "../utils/biz";
 // import { fetchMetadata } from "../utils/request";
 
 export type RSSItemWithStatus = RSSItem & {
-  status: SummarizeStatus;
+  status?: SummarizeStatus;
 };
 
 const MIN_SUMMARIZE_CHARACTER_LIMIT = 100;
@@ -194,15 +194,22 @@ function generateDigestTemplate(provider: Provider, items: RSSItemWithStatus[]):
   return digest;
 }
 
+function addEllipsis(content: string, maxLen: number): string {
+  if (content.length > maxLen) {
+    return content.substring(0, maxLen) + "...";
+  }
+  return content;
+}
+
 // 格式化单个项目以用于摘要
-function formatItemForDigest(item: RSSItemWithStatus, prefixStr?: string): string {
-  return `### ${prefixStr ?? ""}${item.title}  @([${item?.feed?.title}](${addUtmSourceToUrl(
-    item?.feed?.url ?? "",
-  )}))\n${item.coverImage ? `![cover](${item.coverImage})\n\n` : ""}${
-    item.summary || item.content?.substring(0, THRESHOLDS_FOR_TRUNCATION) + "..."
+export function formatItemForDigest(item: RSSItemWithStatus, prefixStr?: string, ignoreFeed?: boolean): string {
+  return `### ${prefixStr ?? ""}${item.title}  ${
+    ignoreFeed ? "" : `@([${item?.feed?.title}](${addUtmSourceToUrl(item?.feed?.url ?? "")}))`
+  }\n${item.coverImage ? `![cover](${item.coverImage})\n\n` : ""}${
+    item.summary || addEllipsis(item.content || "", THRESHOLDS_FOR_TRUNCATION)
   }\n\n[Source Link](${normalizeUrlForMarkdown(item.link ?? "")})\n\n${
     item.status === "summraized" ? `\`✨AI Summarized\`  ` : ""
-  }${["raw", "failedToSummarize"].includes(item.status) ? `\`Raw Content\`  ` : ""}\`Pub Time: ${dayjs(
+  }${["raw", "failedToSummarize"].includes(item?.status ?? "") ? `\`Raw Content\`  ` : ""}\`Pub Time: ${dayjs(
     item.pubDate,
   ).format("YYYY-MM-DD HH:mm")}\`  \`Creator: ${item.creator ?? "none"}\`\n\n`;
 }
@@ -218,7 +225,7 @@ function extractTextFromXML(xml: string): string {
 }
 
 // 格式化RSS Items，对content进行判断，如果是xml，则提取其中text
-async function formatRSSItems(items: RSSItem[]): Promise<RSSItem[]> {
+export async function formatRSSItems(items: RSSItem[]): Promise<RSSItem[]> {
   return Promise.all(
     items.map(async (item) => {
       const content = (item.content || "").replace("&lt;", "<").replace("&gt;", ">");
@@ -335,7 +342,7 @@ export async function genDigest(options: {
   return {
     content: generateDigestTemplate(options.provider, formatedItems),
     items: formatedItems.map((item) => ({
-      status: item.status,
+      status: item.status!,
     })),
     createAt: Date.now(),
   };